+++ packages/storybook/stc.config.ts
@@ -93,7 +93,7 @@ const styles: StclandStyles = {
     root: 'flex  w-100 gap-2 ',
     label: 'text-md text-sky-500',
     inputRoot: {
-      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full dark:[color-scheme:dark]',
+      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full',
       hover: 'hover:border-neutral-stroke-dark',
       pressed: 'active:text-neutral-text-icon-body',
       disabled: 'disabled:text-neutral-stroke-default disabled:bg-neutral-surface-disabled'
@@ -102,7 +102,7 @@ const styles: StclandStyles = {
     root: 'flex  w-100 gap-2 ',
     label: 'text-md text-sky-500',
     inputRoot: {
-      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full dark:[color-scheme:dark]',
+      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full',
       hover: 'hover:border-neutral-stroke-dark',
       pressed: 'active:text-neutral-text-icon-body',
       disabled: 'disabled:text-neutral-stroke-default disabled:bg-neutral-surface-disabled'
@@ -112,7 +112,7 @@ const styles: StclandStyles = {
     root: 'flex  w-100 gap-2 ',
     label: 'text-md text-sky-500',
     inputRoot: {
-      default: 'text-base px-4 py-4 border-sky-200 border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full dark:[color-scheme:dark]',
+      default: 'text-base px-4 py-4 border-sky-200 border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full',
       hover: 'hover:border-neutral-stroke-dark',
       pressed: 'active:text-neutral-text-icon-body',
       disabled: 'disabled:text-neutral-stroke-default disabled:bg-neutral-surface-disabled'
@@ -122,7 +122,7 @@ const styles: StclandStyles = {
     root: 'flex  w-100 gap-2 ',
     label: 'text-md text-sky-500',
     inputRoot: {
-      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full dark:[color-scheme:dark]',
+      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full',
       hover: 'hover:border-neutral-stroke-dark',
       pressed: 'active:text-neutral-text-icon-body',
       disabled: 'disabled:text-neutral-stroke-default disabled:bg-neutral-surface-disabled'
@@ -132,7 +132,7 @@ const styles: StclandStyles = {
     root: 'flex  w-100 gap-2 ',
     label: 'text-md text-sky-500',
     inputRoot: {
-      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full dark:[color-scheme:dark]',
+      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full',
       hover: 'hover:border-neutral-stroke-dark',
       pressed: 'active:text-neutral-text-icon-body',
       disabled: 'disabled:text-neutral-stroke-default disabled:bg-neutral-surface-disabled'
@@ -142,7 +142,7 @@ const styles: StclandStyles = {
     root: 'flex  w-100 gap-2 ',
     label: 'text-md text-sky-500',
     inputRoot: {
-      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full dark:[color-scheme:dark]',
+      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full',
       hover: 'hover:border-neutral-stroke-dark',
       pressed: 'active:text-neutral-text-icon-body',
       disabled: 'disabled:text-neutral-stroke-default disabled:bg-neutral-surface-disabled'